chore!: use u64 for shortint metadata
This switches from usize to u64 for shortint's metdata:
* Degree
* MaxDegree
* CarryModulus
* MessageModulus

The reasoning is that usize should be preferred when the value is used
as some kind of index, memory access, etc, and not numbers like these
metadata are.

This is a breaking API change
This is also a somewhat breaking serialization change
depending on the serialization format (bincode should be ok as it
encodes usize as u64)

+++ tfhe/benches/shortint/bench.rs
@@ -94,7 +94,7 @@ fn bench_server_key_unary_function<F>(
 
         let mut rng = rand::thread_rng();
 
-        let modulus = cks.parameters.message_modulus().0 as u64;
+        let modulus = cks.parameters.message_modulus().0;
 
         let clear_text = rng.gen::<u64>() % modulus;
 
@@ -138,7 +138,7 @@ fn bench_server_key_binary_function<F>(
 
         let mut rng = rand::thread_rng();
 
-        let modulus = cks.parameters.message_modulus().0 as u64;
+        let modulus = cks.parameters.message_modulus().0;
 
         let clear_0 = rng.gen::<u64>() % modulus;
         let clear_1 = rng.gen::<u64>() % modulus;
@@ -184,7 +184,7 @@ fn bench_server_key_binary_scalar_function<F>(
 
         let mut rng = rand::thread_rng();
 
-        let modulus = cks.parameters.message_modulus().0 as u64;
+        let modulus = cks.parameters.message_modulus().0;
 
         let clear_0 = rng.gen::<u64>() % modulus;
         let clear_1 = rng.gen::<u64>() % modulus;
@@ -229,7 +229,7 @@ fn bench_server_key_binary_scalar_division_function<F>(
 
         let mut rng = rand::thread_rng();
 
-        let modulus = cks.parameters.message_modulus().0 as u64;
+        let modulus = cks.parameters.message_modulus().0;
         assert_ne!(modulus, 1);
 
         let clear_0 = rng.gen::<u64>() % modulus;
@@ -270,7 +270,7 @@ fn carry_extract_bench(c: &mut Criterion, params_set: BenchParamsSet) {
 
         let mut rng = rand::thread_rng();
 
-        let modulus = cks.parameters.message_modulus().0 as u64;
+        let modulus = cks.parameters.message_modulus().0;
 
         let clear_0 = rng.gen::<u64>() % modulus;
 
@@ -306,7 +306,7 @@ fn programmable_bootstrapping_bench(c: &mut Criterion, params_set: BenchParamsSe
 
         let mut rng = rand::thread_rng();
 
-        let modulus = cks.parameters.message_modulus().0 as u64;
+        let modulus = cks.parameters.message_modulus().0;
 
         let acc = sks.generate_lookup_table(|x| x);
