+++ tfhe/src/integer/server_key/radix_parallel/modulus_switch_compression.rs
@@ -66,7 +66,7 @@ impl ServerKey {
             "Compression does not support message_modulus > carry_modulus"
         );
         assert!(
-            self.key.max_noise_level.get() >= self.message_modulus().0 as u64 + 1,
+            self.key.max_noise_level.get() >= self.message_modulus().0 + 1,
             "Compression does not support max_noise_level < message_modulus + 1"
         );
 
@@ -113,11 +113,11 @@ impl ServerKey {
     ) -> Vec<Ciphertext> {
         let message_extract = self
             .key
-            .generate_lookup_table(|x| x % self.message_modulus().0 as u64);
+            .generate_lookup_table(|x| x % self.message_modulus().0);
 
         let carry_extract = self
             .key
-            .generate_lookup_table(|x| x / self.message_modulus().0 as u64);
+            .generate_lookup_table(|x| x / self.message_modulus().0);
 
         let mut blocks: Vec<Ciphertext> = compressed_ct
             .paired_blocks
