+++ tfhe/src/integer/server_key/radix_parallel/tests_unsigned/test_sum.rs
@@ -48,7 +48,7 @@ where
     let mut rng = rand::thread_rng();
 
     // message_modulus^vec_length
-    let modulus = cks.parameters().message_modulus().0.pow(NB_CTXT as u32) as u64;
+    let modulus = cks.parameters().message_modulus().0.pow(NB_CTXT as u32);
 
     for len in [1, 2, 15, 16, 17, 64, 65] {
         for _ in 0..nb_tests_smaller {
@@ -157,8 +157,7 @@ where
         .parameters()
         .message_modulus()
         .0
-        .pow(crate::integer::server_key::radix_parallel::tests_unsigned::NB_CTXT as u32)
-        as u64;
+        .pow(crate::integer::server_key::radix_parallel::tests_unsigned::NB_CTXT as u32);
 
     executor.setup(&cks, sks);
 
@@ -195,7 +194,7 @@ where
     let mut rng = rand::thread_rng();
 
     // message_modulus^vec_length
-    let modulus = cks.parameters().message_modulus().0.pow(NB_CTXT as u32) as u64;
+    let modulus = cks.parameters().message_modulus().0.pow(NB_CTXT as u32);
 
     for len in [1, 2, 15, 16, 17, 64, 65] {
         for _ in 0..nb_tests_smaller {
